Add SettingsModule and add guild rank roles to GuildSettings
Note:
I've decided to implement settings manually (instead of automatically via reflection), because this way I have more control over the settings and I can use the default type readers.
The only negative point is that I need to keep this updated and there's gonna be a lot of copy paste, but I think it's worth it.