Make LiteDatabase singleton
According to the developer, if the threads are in the same process then we should just reuse the same LiteDatabase instance.
Source: https://github.com/mbdavid/LiteDB/issues/725

For now I'm trying to keep the repositories as transient and use the Database.Checkpoint method, which should commit the changes to the file.
If this doesn't work well then I'll either change the repositories to singleton or I'm just gonna remove the disposable pattern.