Fix repository and database disposing
Repository did not implement sync Dispose and IServiceScope cannot use the async one.

Database disposing is a workaround because the async version of LiteDatabase creates background thread which blocks the application from closing if not disposed.
It's not correct to reference LiteDB here, but there wasn't any easy and simple solution for that (at least what I've tried).
So I'm gonna fix this when we introduce domain events.